+++ R/Senn2013.R
@@ -59,9 +59,9 @@
 #' 
 #' # Fixed effects model
 #' #
-#' net1 <- netmeta(TE, seTE, treat1, treat2,
+#' net1 <- netmeta(TE, seTE, treat1.long, treat2.long,
 #'                 studlab, data = Senn2013,
-#'                 comb.random = FALSE)
+#'                 comb.random = FALSE, nchar.trts = 4)
 #' net1
 #' net1$Q.decomp
 #' 
@@ -78,8 +78,8 @@
 #' 
 #' # Random effects model
 #' #
-#' net2 <- netmeta(TE, seTE, treat1, treat2, studlab, data = Senn2013,
-#'                 comb.fixed = FALSE)
+#' net2 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
+#'                 data = Senn2013, comb.fixed = FALSE)
 #' net2
 #' forest(net2, ref = "plac")
 #' }