+++ R/print.decomp.design.R
@@ -49,7 +49,7 @@ print.decomp.design <- function(x,
                                 digits.tau2 = gs("digits.tau2"),
                                 scientific.pval = gs("scientific.pval"),
                                 big.mark = gs("big.mark"),
-                                nchar.trts = x$x$nchar.trts,
+                                nchar.trts = x$nchar.trts,
                                 legend = TRUE,
                                 ...) {